temp fix for supabase errors
