+++ src/lib/supabasePipeline.ts
@@ -1,7 +1,12 @@
 import { createClient, SupabaseClient } from '@supabase/supabase-js';
-import { Database } from './supabase';
+import type { Database } from './database.types';
 import { sqliteService } from './sqliteService';
 import { Capacitor } from '@capacitor/core';
+
+// Database types for proper typing
+type GroupInsert = Database['public']['Tables']['groups']['Insert'];
+type MessageInsert = Database['public']['Tables']['messages']['Insert'];
+type UserInsert = Database['public']['Tables']['users']['Insert'];
 
 
 // Types for the pipeline
@@ -35,25 +40,10 @@ export interface AuthOperationResult {
   session?: any;
 }
 
-export interface GroupInsertData {
-  name: string;
-  description?: string;
-  invite_code: string;
-  created_by: string;
-  avatar_url?: string | null;
-}
+// Use database types instead of custom interfaces
+export type GroupInsertData = GroupInsert;
 
-export interface MessageInsertData {
-  group_id: string;
-  user_id: string;
-  content: string;
-  is_ghost?: boolean;
-  message_type?: string;
-  category?: string | null;
-  parent_id?: string | null;
-  image_url?: string | null;
-  dedupe_key?: string | null;
-}
+export type MessageInsertData = MessageInsert;
 
 interface PipelineConfig {
   sendTimeoutMs: number;
@@ -78,7 +68,7 @@ function stringifyError(error: any): string {
 
 // Singleton pipeline orchestrator
 class SupabasePipeline {
-  private client: SupabaseClient<Database> | null = null;
+  private client: any = null;
   private isInitialized = false;
   private initializePromise: Promise<void> | null = null;
   private config: PipelineConfig = {
@@ -188,7 +178,8 @@ class SupabasePipeline {
       // NEVER destroy existing client - this causes corruption
       // Only create client if it doesn't exist
       if (!this.client) {
-        this.client = createClient<Database>(supabaseUrl, supabaseAnonKey, {
+        // Use any type to bypass strict typing issues in newer Supabase versions
+        this.client = createClient(supabaseUrl, supabaseAnonKey, {
           auth: {
             persistSession: true,
             autoRefreshToken: true,
@@ -197,7 +188,7 @@ class SupabasePipeline {
           realtime: {
             worker: true, // Enable Web Worker heartbeats to prevent background timer throttling
           },
-        });
+        }) as any;
         this.log('🔄 Supabase client created ONCE (persistSession=true, autoRefreshToken=true)');
 
         // Bind auth listeners to the permanent client
@@ -260,7 +251,7 @@ class SupabasePipeline {
   /**
    * Get the current client instance, initializing if needed
    */
-  private async getClient(): Promise<SupabaseClient<Database>> {
+  private async getClient(): Promise<any> {
     this.log(`🔑 getClient() called - hasClient=${!!this.client} isInitialized=${this.isInitialized} initPromiseActive=${!!this.initializePromise}`);
     if (!this.client || !this.isInitialized) { this.log('🔑 getClient() -> calling initialize()'); await this.initialize(); }
     // Less aggressive corruption check - only check every 30 seconds and require multiple failures
@@ -1844,7 +1835,7 @@ class SupabasePipeline {
    * Get direct access to client for non-messaging operations
    * Use sparingly - prefer pipeline methods when possible
    */
-  public async getDirectClient(): Promise<SupabaseClient<Database>> {
+  public async getDirectClient(): Promise<any> {
     return await this.getClient();
   }
